BF: imports and usage relative to root of repository, not to main.py
The usage did error for me because of the format --web.listen-address
seems to expect -- I assume that comment was out of date.

The relative import did only work when I then run the project from the
directory in which main.py is, not from the root of the repository.